requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.8.2
python-docx==1.1.2
pdfminer.six==20231228
rapidfuzz==3.9.6
openai==1.52.2
python-multipart==0.0.9
appdirs==1.4.4
